+++ core/src/main/java/cucumber/runtime/formatter/JUnitFormatter.java
@@ -16,6 +16,7 @@ import gherkin.formatter.model.Step;
 import org.w3c.dom.Document;
 import org.w3c.dom.Element;
 import org.w3c.dom.Node;
+import org.w3c.dom.NodeList;
 
 import javax.xml.parsers.DocumentBuilderFactory;
 import javax.xml.parsers.ParserConfigurationException;
@@ -95,8 +96,11 @@ class JUnitFormatter implements Formatter, Reporter, StrictAware {
     @Override
     public void done() {
         try {
-            //set up a transformer
+            // set up a transformer
+            rootElement.setAttribute("name", JUnitFormatter.class.getName());
             rootElement.setAttribute("failures", String.valueOf(rootElement.getElementsByTagName("failure").getLength()));
+            rootElement.setAttribute("skipped", String.valueOf(rootElement.getElementsByTagName("skipped").getLength()));
+            rootElement.setAttribute("time", sumTimes(rootElement.getElementsByTagName("testcase")));
             if (rootElement.getElementsByTagName("testcase").getLength() == 0) {
                 addDummyTestCase(); // to avoid failed Jenkins jobs
             }
@@ -145,6 +149,24 @@ class JUnitFormatter implements Formatter, Reporter, StrictAware {
     private void handleHook(Result result) {
         testCase.hookResults.add(result);
         testCase.updateElement(doc, root);
+    }
+
+    private String sumTimes(NodeList testCaseNodes) {
+        double totalDurationSecondsForAllTimes = 0.0d;
+        for( int i = 0; i < testCaseNodes.getLength(); i++ ) {
+            try {
+                double testCaseTime =
+                        Double.parseDouble(testCaseNodes.item(i).getAttributes().getNamedItem("time").getNodeValue());
+                totalDurationSecondsForAllTimes += testCaseTime;
+            } catch ( NumberFormatException e ) {
+                throw new CucumberException(e);
+            } catch ( NullPointerException e ) {
+                throw new CucumberException(e);
+            }
+        }
+        DecimalFormat nfmt = (DecimalFormat) NumberFormat.getNumberInstance(Locale.US);
+        nfmt.applyPattern("0.######");
+        return nfmt.format(totalDurationSecondsForAllTimes);
     }
 
     private void increaseAttributeValue(Element element, String attribute) {